requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
celery
redis
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
Pillow==9.3.0
